Some changes in the headers and annotations.
A new plugin has been added : it is now possible to add all the sources
of on type (shp for instance) that are contained under a given
directory. The folders are analyzed recursively. This plugin is directly
obtained from the one which is used to import files one by one.