Fixing a unmanaged NullPointerException.
When trying to get the spatial field index in a SpatialDataSourceDecorator without spatial information, we obtained a null pointer Exception : getSpatialFieldIndex returned -1. When trying to read informations at index -1, other methods threw a NullPointerException. 

getSpatialFieldIndex now perform a new check, and throw a DirverException (which is cleaner, and with a message the user can understand) if it does not find the spatial index.