Fix the behaviour of many methods and classes, that were relying on the old GDMS type system. Hopefully, I won't break much things ...
I say hopefully, mainly because of the direct dependency of many plugins and tools on the correct management  of geometry type constraints. I've tried to be as accurate as possible in all the cases, using generic collections with planar dimension constraints when possible. However, I can't guarantee the correct behaviour of the software without a really deep testing of it. And I'm just about to start these tets... Shall the force be with us.