FIX #553 A test is done in selectequalplugin to support empty value
Task #255
Add in tablecomponent the possibility to revert a selection


---
Also : 
Improve the use of findreplacedialog (no duplication, add crtl+ h to open the dialog)
Improve the search tool bar in an attributes table : 
SQL query is not yet supported, only simple pattern matches. 
SQL query will be added with the new filterdatasourcedecorator.