Preparing a merge with the trunk. It will certainly occur later, but, well... at least, we're almost ready
I've made it in a separate branch rather than directly in the trunk for
many reasons. The main one is that I don't want to break the history of
our project. Making it this way, we'll just patch the trunk with the
result obtained here, introducing just the needed modifications.

It is clearer than a merge. Indeed, we won't be stuck (or rather, it
will be less complicated) because of conflicts. And there will be some
logic : we will "just" replace the renderer, and patch the UI as
needed...

As you can see, the commit is very, very huge. I've imported all the
new and modified files that were in the trunk and not in this branch.
For some of them, the history will be completely broken... in this
branch. As explained above, the goal is not to break the trunk's
history. And as we will try to merge this branch in the trunk... It
should work. I hope so...

I've finally split the translation keys. I forgot them when seprating
core en ui in this branch... even though it had been made properly in
the trunk. Shame on me.

I'm not sure the diff is worth reading. Indeed, it's mainly made of :

- changes to use the newest API of GDMS2 (DataSource and co)
- key refactoring
- Types conformance.
- ...

Thanks to this commit, it's now possible to use the latest beanshell and
SQL console, and the latest GDMS implementation of the SQL language. We
still suffer from the frightening interface to build maps, but there is
still some code to write in order to be able to fix it... BTW, I think
such an interface can have a meaning in the UI we will release with the
final 4.0. And we have all the cool edition tools too - not tested them
yet, they may be broken.

that's all. Given the size of the commit, it is, really, a very short
description ;-)