We were doing an ugly hack to "support" FilterEncogind in the Rule's Filter element. However, this hack could absolutely not stay as is. Moreover, according to the current FES and SE norms, we are not forced to support FES in this place - we can just say that we can't do it. Finally, because of the recent optimizations that have been made on GDMS and GDMSQL, we definitely can't use FilterDataSourceDecorator instances as we did in the renderer.
Because of all these reasons, the rendering strategy has changed. We
will draw all the features that are visible in the displayed extent,
whatever the content of the Filter markup is. To do so, we now directly
use the spatial index of the DataSource, and not a FilterDataSourceDecorator
anymore. We are now able again to draw all the needed geometries. And
there is a performance improvement for the rendering in this OrbisGIS version,
too.