updated antlr to version 3.4.
This was not possible before because of a bug in 3.4 that generated invalid java sources.
The bug is probably still there in antlr, but for some reason we do not trigger it anymore, so I see no reason not to upgrade.
Downgrading to 3.3 is always possible if another grammar change triggers it again.

Note that this does not change the dependency problem with antlr: it is still not properly bootstrapped, so we still depends on antlr 2.7.7.