improve type inference for NULL used with operators.
Some more information on this: all null values have a type, but
since SQL is statically typed and we want user to use NULL
without always specifiying its type, we need some special type
inference to get the result type of the NULL.

I made the choice of letting NULL keep its declared Type.NULL type
until the end, except when it is the root of the expression tree:
 * if NULL is the child of an operator, it keeps its type but we
attempt to infere the type of the result of the operator
evaluation;
 * if NULL is directly the root (and single element) of an
expression tree, then its type will be changed. This includes
function arguments and all expressions used in SQL (select list
elements, where filters, ...). This will be handled in another
commit.

This commit makes the result of "value operator NULL" being of the
same type as "value", except obviously when the operator has a
fixed return type. This lets some unary operator forward the NULL
type higher in the tree.