fix a laziness problem in Command.doWork (new with scala 2.10.0-M4)
Some commands used 'doWork' instead of 'execute' sometimes, i.e.
they did not need laziness of the RowStream from their children
since they did not use these anyway. This was not a problem before
because the call to toIterator on a view in Command.execute forced
the view to be executed anyway.

Now the iterator of a view seems to be really lazy, so the children
were never actually executed.