Replace calls to the pair (DataSource,long) with calls to a HashMap.
The main goal here is to gathered the calls to the data in the
DataSource in one place and to perform them as soon as possible. With
the dependsOnFeature method, the first was already here.

This strategy will be particularly useful for complex types where many
analysis are made on the same field of the DataSource. As the data will
be stored in the Map and already in memory, we will gain some calls and
some disk access (in the best case).

This strategy change was first intended to improve the point cache
performance. It did not work, but I think it is useful anyway.