Change the rendering strategy.
Levels are not in use anymore. From now on, we draw symbolizers in the order they
are found in the rules. We currently simply use indices of the generated
list of symbolizers (which is not sorted anymore in Style). It should be
enough : this list is generated each time the time is rendered.