Improve management of dash arrays in the simple editor. Fix #62
Wrong inputs are now better managed : random text and dash arrays with
negative values are removed when focus is lost or when the style is
created by clicking OK. These invalid patterns are replaced with an
empty string.

This won't prevent the use of bad patterns in external files. However,
checks are made to let the user change these bad values through the
simple interface (when possible).