Changes the checks that are made when a layer is removed.
When removing a layer from the TOC, we will now check whether
there is still a layer with spatial data in the layer model. If there is
not (ie if we have only layer collections), the extents of the
MapControl and of the MapContext are set to null. This way, when adding
back a layer with spatial data, the extent of the map control is
modified to match the new bounding box of the layer model.

As for PR #83, this commit is open for discussion. Indeed, I change the
behaviour of listeners here, and consequently modify their side
effects... Moreover, I'm absolutely not able to determine what
consequences such a modification could have on the WMS project...