Adds a checkbox in the UI for symbolizers where the line is optional.
We use the enable property of the Swing containers to do this.
PnlUniqueLineSE and its children have a new method (isLineOptional) that
is used to add the checkbox when needed (unique line MUST have a line
configuration)..

This commit adds also a memory for lines that was set but disabled.
Indeed, the actual legend is replaced with a NullPenStrokeLegend in the
analysis, while the former actual analysis is kept. This way, we can
enable it back as is on demand.