Short version : adds a default constructor in ProportionalPoint
Longer version : this has many consequences in many places, of course.
Particularly, we need default constructors in many places from now on.
Moreover, types for legends on well-known names have been made more
specific to improve interactions with them.

This new constructor is tested... using the analyzer on PointSymbolizer.
That seemed to be both an efficient and straightforward strategy.