add support for SQL parametrized statements + scripts.
Parameters are of the form @{paramname} and can be used anywhere
on expects:
 - a constant value ("SELECT @{param} AS toto;")
 - a field name ("SELECT @{field} FROM hello;")
 - a table name ("SELECT * FROM @{table}")

Both SQLStatement and SQLScript can receive as many parameters
as necessary.

An SQLStatement could already be persisted to disk, but not a
whole script.
This become now as easy as

val a = Engine.parseScript("<some sql>")
// or Engine.parseScript(new File("/tmp/f.sql")
a.save(new FileOutputStream("/tmp/f.bsql"))

and later

val b = Engine.loadScript(new FileInputStream("/tmp/f.bsql"))
b.execute

Same thing can be done for individual statements with
Engine.parse(...) and Engine.load(...)