Fix a problem with the GeoJSON parser. There was a problem whithe the management of metadata. The original sorted map was not kept but replace by a HahsMap. Then, we were betting on the fact the insertion order of the HashMap would be kept. It is not. It is expected not to be. And that caused the bug : Metadata were not presented altered the right way.
I've added a test case for this. The input file has been produced from a
data file that were correctly exported to JSON, but not correctly
imported back. Metadata were misplaced on the columns... Should be fixed
now. Even if I'm afraid we may face other similar, but more complicated,
problems in this piece of code.