Fix a test linked to the shapefile driver.
We had a problem in the shapefile reader with the last JTS release.
Indeed, we were not forcing the dimension to 3 so that we couldn't
change the dimension later. Dimension management in JTS is still not
perfect : we currently can't change dimension during runtime. Anyway, it
should work for us now.