Fix slow scrolling!
Fixes #440.

It's interesting to note that when I used setUnitIncrement(int) for only
the vertical scrollbar of the JScrollPanel holding a legend (symbol)
panel, the scrolling remained incredibly slow. To be safe, I introduce a
getJScrollPane(ISELegendPanel) method where I set both the vertical and
horizontal unit increments to the constant INCREMENT (=16). This fixes
the problem.