Changes the representation of double values in interval calssifications. I'm playing with significant digits here... For values greater than one, I don't want more than 4 numbers in the fraction part of the number. For values lower than one, I want to be able to display things with at least for significant numbers...
I use a trick here to be able to do that. For numbers whose absolute
value is lower than 1, I compute an approximation that keeps exactly
four digits and I display it normally. When editing it, the complete
value comes back. We loose accuracy for representation in the table but
the does not change things for the accuracy of the symbol
configuration...

In order to validate user input and to be able to parse inputs more
accurately, I use a JFormattedTextField for the edition.