Redo LegendUIChooser.
In the previous commit, I narrowed down the slow initialization problem
to the LegendUIChooser. Here I fix it.

Before, the LegendUIChooser initialized ILegendPanels for all possible
Legends. This was pointless and slowed things down.

Now we pass the SimpleStyleEditor to the constructor, from which we can
recover the currently edited layer. From the layer, we populate an array
of Strings containing only the names of the possible legends. When the
user selects a legend and clicks OK, *only then* is one (and only one)
new ILegendPanel instantiated (using the new ILegendPanelFactory). This
makes things quite a bit snappier.

With these changes, the LegendTree.addLegend() method has become
significantly simpler.

I also included all the missing preview images for interval
classifications (thanks @gpetit).