Translate analysis types.
Note 1: For (Area|Line|Point) Symbolizers, I was able to translate in the
Advanced Style Editor when adding a symbol, but I wasn't able to
translate the default symbol in the Simple Style Editor or the Advanced
Style Editor. I'm not sure why. Maybe these strings are defined in .se
files somewhere, where they can't be translated?

Note 2: There are a lot of additions since apparently the translation
keys for legend and orbisgis-core had never been generated.