Remove usages of PnlRule#initialize(LegendContext)
First step in major refactor of Pnl constructors. We want to remove all
usages of initialize(LegendContext) and instead pass a LegendContext to
the constructor. Separating the initialization from the constructor is
one of the main reasons for #484.

We also make RuleWrapper take a LegendContext in the constructor, and
remove two unused constructors.