related to #523.  We were not able to open the simple editor after having saved a unique symbol area with no stroke. Indeed, the fillLegendMemory attached to LinePanel was a NullPenStrokeLegend so the init method threw a IllegalStateException...
The fix is not that clean. We have code where constructors have coupling
with listeners and that produce border effects. I've decided to go as
fast as I could as this design has already been discussed and accepted.
It may change later, however, in order to separate behaviour management
from initialization.

I rely here on a new field that lets me know if the stroke (or fill,
resp.) was originally set so that I can decide if fields must be enabled
or not. It seems to be working, bu I may have missed something, so please
test it.